criando botão para sair do jogo + controlando encerrar sessão da game board

+++ components/MainHeader.tsx
@@ -15,33 +15,37 @@ import { toast } from "sonner";
 import Image from "next/image";
 import Logo from "@/assets/icons/memory-game-logo.svg";
 import { socket } from "@/app/socket";
+import DialogConfirmExitRoom from "@/app/room/[id]/components/DialogConfirmExitRoom";
 
-export default function MainHeader({ showLogo }: { showLogo?: boolean }) {
+export default function MainHeader({
+  showLogo,
+  isGameBoard,
+}: {
+  showLogo?: boolean;
+  isGameBoard?: boolean;
+}) {
   const router = useRouter();
   const [nickName, setNickName] = useState<string>("");
   const [avatar, setAvatar] = useState<string>("");
   const [loading, setLoading] = useState<boolean>(false);
   const pathname = usePathname();
   const { id } = useParams();
+  const [dialogConfirmExitRoom, setDialogConfirmExitRoom] =
+    useState<boolean>(false);
+  const [dialogIsLoggedOut, setDialogIsLoggedOut] = useState<boolean>(false);
 
   function handleLogout() {
-    if (pathname.includes("/room/")) {
-      const user = getUserLocal();
+    if (!isGameBoard) {
+      localStorage.clear();
+      router.push("/auth");
 
-      if (user && id) {
-        socket.emit("requestUserLoggedOut", {
-          playerId: user.id,
-          roomId: Number(id),
-        });
-      }
+      toast.info("Você foi desconectado. Faça login novamente.", {
+        duration: 3000,
+      });
+    } else {
+      setDialogIsLoggedOut(true);
+      setDialogConfirmExitRoom(true);
     }
-
-    localStorage.clear();
-    router.push("/auth");
-
-    toast.info("Você foi desconectado. Faça login novamente.", {
-      duration: 3000,
-    });
   }
 
   useEffect(() => {
@@ -55,7 +59,7 @@ export default function MainHeader({ showLogo }: { showLogo?: boolean }) {
   }, []);
 
   return (
-    <div className="w-full h-24 bg-gray-800 border-b-solid border-b-2 border-b-purple-800">
+    <div className="w-full h-28 sm:h-24 bg-gray-800 border-b-solid border-b-2 border-b-purple-800">
       <div className="flex items-center justify-between h-full px-5">
         {showLogo && (
           <div>
@@ -67,7 +71,22 @@ export default function MainHeader({ showLogo }: { showLogo?: boolean }) {
           </div>
         )}
 
-        <div>
+        <div className="flex flex-col sm:flex-row items-center gap-5">
+          <div>
+            {isGameBoard && (
+              <div>
+                <Button
+                  onClick={() => {
+                    setDialogConfirmExitRoom(true);
+                    setDialogIsLoggedOut(false);
+                  }}
+                  className="bg-red-600 hover:bg-red-800 transition-all text-gray-50"
+                >
+                  Sair do jogo
+                </Button>
+              </div>
+            )}
+          </div>
           <Popover>
             <PopoverTrigger asChild>
               <Button className="bg-transparent border border-solid border-purple-600 flex items-center gap-2">
@@ -97,6 +116,15 @@ export default function MainHeader({ showLogo }: { showLogo?: boolean }) {
           </Popover>
         </div>
       </div>
+
+      <DialogConfirmExitRoom
+        onClose={() => {
+          setDialogConfirmExitRoom(false);
+          setDialogIsLoggedOut(false);
+        }}
+        open={dialogConfirmExitRoom}
+        loggedOut={dialogIsLoggedOut}
+      />
     </div>
   );
 }
